proyecto/notificacion-service/requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
kafka-python==2.0.2
